tp remise /post /put /del /get + error handling

+++ src/main/java/org/example/tpremise/mappers/RemiseMapper.java
@@ -10,7 +10,7 @@ public class RemiseMapper implements RowMapper<Remise> {
     @Override
     public Remise mapRow(ResultSet rs, int rowNum) throws SQLException {
             Remise remise = new Remise();
-            remise.setId(rs.getLong("id"));
+            remise.setId(rs.getInt("id"));
             remise.setMontantMin(rs.getDouble("montant_min"));
             remise.setMontantMax(rs.getDouble("montant_max"));
             remise.setTaux(rs.getDouble("taux"));
