+++ src/main/java/org/example/tpremise/mappers/TransactionMapper.java
@@ -11,7 +11,7 @@ public class TransactionMapper implements RowMapper<Transaction> {
     @Override
     public Transaction mapRow(ResultSet rs, int rowNum) throws SQLException {
         Transaction transaction = new Transaction();
-        transaction.setId(rs.getLong("id"));
+        transaction.setId(rs.getInt("id"));
         transaction.setDate(rs.getTimestamp("date").toLocalDateTime());
         transaction.setMontantAvant(rs.getDouble("montant_avant"));
         transaction.setMontantApres(rs.getDouble("montant_apres"));
